Move common logic from interop to impl package
