+++ pekko-classic/src/main/scala/net/ruippeixotog/akka/testkit/specs2/PekkoMatchers.scala
@@ -4,7 +4,7 @@ import scala.concurrent.duration.FiniteDuration
 
 import org.apache.pekko.testkit.TestKitBase
 
-import net.ruippeixotog.akka.testkit.specs2.interop.ClassicMatchers
+import net.ruippeixotog.akka.testkit.specs2.impl.ClassicMatchers
 
 trait PekkoMatchers extends ClassicMatchers[TestKitBase] {
   protected def defaultReceiveTimeout(testkit: TestKitBase): FiniteDuration =
